+++ Runtime/PlayerAccount/ProfilePictureSize.cs
@@ -1,0 +1,9 @@
+namespace Agava.YandexGames
+{
+    public enum ProfilePictureSize
+    {
+        small,
+        medium,
+        large
+    }
+}